Refactor Bar into separate file
